+ Log/Linear histogram + Autosave options if modified + Option to show/hide over-/under-explosure + Palette magnify + 'double' convertion function + Changed using of math functions to built-in + Histogram auto ajust + Histogram ignore boundaries * A lot of UI improvements * etc.
git-svn-id: https://tools.assembla.com/svn/chdk/branches/grand@60 6794e30b-3f2a-0410-a806-a2bbca1c07ff

+++ platform/a610/main.c
@@ -1,6 +1,7 @@
 #include "platform.h"
 #include "core.h"
 #include "stdlib.h"
+#include "math.h"
 #include "keyboard.h"
 
 /* VxWorks stuff */
@@ -571,7 +572,7 @@ int shooting_get_av()
 }
 
 int shooting_get_real_av() {
-    return (int)(((float)powf(1.4142135623730950488016887242097/* sqrt(2) */, ((float)GetCurrentAvValue())/96.0))*100.0);
+    return (int)(((double)pow(1.4142135623730950488016887242097/* sqrt(2) */, ((double)GetCurrentAvValue())/96.0))*100.0);
 }
 
 void shooting_set_av(int v)
